requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
bs4==0.0.1
Django==1.10.6
django-auth-ldap==1.2.10
django-taggit==0.22.0
django-widget-tweaks==1.4.1
djangorestframework==3.6.2
packaging==16.8
pyparsing==2.2.0
python-ldap==2.4.32
six==1.10.0
